+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- VIMEET-2526 - Happening question sort by like or sort by date
+
 ### Fixed
 - Avoid error if country is not defined when displaying participant info
 - MV-333 - Impossible to use Google & Linkedin login